# Implement deepCopy as copying, instead of merge