--- a/protobuf/lib/src/protobuf/pb_map.dart
+++ b/protobuf/lib/src/protobuf/pb_map.dart
@@ -131,4 +131,8 @@ class PbMap<K, V> extends MapBase<K, V> {
     }
     return this;
   }
+
+  PbMap<K, V> deepCopy() {
+    throw 'TODO';
+  }
 }