--- a/protoc_plugin/lib/src/message_generator.dart
+++ b/protoc_plugin/lib/src/message_generator.dart
@@ -474,7 +474,7 @@ class MessageGenerator extends ProtobufContainer {
           "'See https://github.com/google/protobuf.dart/issues/998.')",
         );
         out.println(
-          '$classname clone() => $classname()..mergeFromMessage(this);',
+          '$classname clone() => deepCopy();',
         );
         out.println(
           '@$coreImportPrefix.Deprecated('