… benchmarks/bin/deep_copy.dart
… protobuf/CHANGELOG.md
… protoc_plugin/CHANGELOG.md
… protoc_plugin/lib/src/gen/google/protobuf/descriptor.pb.dart
… protoc_plugin/lib/src/gen/google/protobuf/duration.pb.dart
… protoc_plugin/test/generated_message_test.dart
… protoc_plugin/test/goldens/deprecations.pb.dart
… protoc_plugin/test/goldens/doc_comments.pb.dart
… protoc_plugin/test/goldens/grpc_service.pb.dart
… protoc_plugin/test/goldens/imports.pb.dart
… protoc_plugin/test/goldens/int64.pb.dart
… protoc_plugin/test/goldens/messageGenerator.pb.dart
… protoc_plugin/test/goldens/messageGenerator.pb.dart.meta
… protoc_plugin/test/goldens/oneMessage.pb.dart
… protoc_plugin/test/goldens/oneMessage.pb.dart.meta
… protoc_plugin/test/goldens/service.pb.dart
… protoc_plugin/test/oneof_test.dart