@osa1 2022-09-06
This causes a lot of breakage internally. I think one of the reasons is probably that I'm assuming `PbList` for repeated fields and `PbMap` for map fields, but map and list allocation is done via `GeneratedMessage.createRepeatedField` (returns `List<T>`) and `GeneratedMessage.createMapField` (returns `Map<K, V>`). I should check the field type (instead of field value type), and use `Map` and `List` API similar to how we merge map and list fields.

Since we can't have nested lists and maps, cloning of nested maps and lists will have to go through `GeneratedMessage.deepCopy`, which will again look at field types and use List/Map API to clone maps and lists. So we won't need `deepCopy` in `PbMap` and `PbList` types.

Interestingly, we assume `PbList` and `PbMap` in `_FieldSet._markReadOnly`:

```dart
void _markReadOnly() {
  if (_isReadOnly) return;
  _frozenState = true;
  for (var field in _meta.sortedByTag) {
    if (field.isRepeated) {
      PbList? list = _values[field.index!];
      if (list == null) continue;
      list.freeze();
    } else if (field.isMapField) {
      PbMap? map = _values[field.index!];
      if (map == null) continue;
      map.freeze();
    } else if (field.isGroupOrMessage) {
      final entry = _values[field.index!];
      if (entry != null) {
        (entry as GeneratedMessage).freeze();
      }
    }
  }

  _extensions?._markReadOnly();
  _unknownFields?._markReadOnly();
}
```

If I'm right that the breakage here caused by users overriding list and map field types, then I'm not sure how `_markReadOnly` works. Maybe users that override map and list types never freeze their messages?

We should just rewrite this library..