@osa1 2022-09-08
I managed to fix at least some of the downstream breakage by removing the assumptions in the implementation about repeated fields being `PbList` and map fields being `PbMap`.

The fact that we csan't assume this means freezing messages is buggy, for the reasons described in my previous comment above.

Performance is still much better than merge-based implementation. I will update the PR description, and refactor the code a little bit (document, revert some of the changes).