@osa1 2025-07-28
We should revive this as it improves cloning performance quite a bit (50% according to my original report in PR description).

As far as I understand from the discussion above the main issue was repeated and map fields were overridden by message classes, but we've since removed those members (96d95225e1843427728572647393b83808534577) so this should probably be good now.

It might make sense to just implement this again from scratch instead of fixing this PR.